+++ astropy/io/ascii/__init__.py
@@ -6,6 +6,7 @@
 from __future__ import absolute_import, division, print_function
 
 from .core import (InconsistentTableError,
+                   ParameterError,
                    NoType, StrType, NumType, FloatType, IntType, AllType,
                    Column,
                    BaseInputter, ContinuationLinesInputter,
@@ -23,6 +24,12 @@ from .basic import (Basic, BasicHeader, BasicData,
                     Tab,
                     NoHeader,
                     CommentedHeader)
+from .fastbasic import (FastBasic,
+                        FastCsv,
+                        FastTab,
+                        FastNoHeader,
+                        FastCommentedHeader,
+                        FastRdb)
 from .cds import Cds
 from .latex import Latex, AASTex, latexdicts
 from .html import HTML