+++ astropy/io/ascii/tests/test_read.py
@@ -15,11 +15,13 @@ from distutils import version
 
 from .common import (raises, assert_equal, assert_almost_equal,
                      assert_true, setup_function, teardown_function)
+from .. import core
 from ....tests.helper import pytest
 
 _NUMPY_VERSION = version.LooseVersion(np.__version__)
 
-def test_convert_overflow():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_convert_overflow(fast_reader):
     """
     Test reading an extremely large integer, which falls through to
     string due to an overflow error (#2234).
@@ -29,7 +31,8 @@ def test_convert_overflow():
     # it falls through to float, so we just accept this as a known issue for
     # numpy < 1.6.
     expected_kind = ('f',) if _NUMPY_VERSION < version.LooseVersion('1.6') else ('S', 'U')
-    dat = ascii.read(['a', '1' * 10000], format='basic', guess=False)
+    dat = ascii.read(['a', '1' * 10000], format='basic',
+                     fast_reader=fast_reader, guess=False)
     assert dat['a'].dtype.kind in expected_kind
 
 
@@ -89,15 +92,17 @@ def test_guess_with_format_arg():
     assert dat.colnames == ['a', 'b']
 
 
-@raises(ValueError)
-def test_read_with_names_arg():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_read_with_names_arg(fast_reader):
     """
     Test that a bad value of `names` raises an exception.
     """
-    dat = ascii.read(['c d', 'e f'], names=('a', ), guess=False)
+    with pytest.raises(ValueError):
+        dat = ascii.read(['c d', 'e f'], names=('a', ), guess=False, fast_reader=fast_reader)
 
 
-def test_read_all_files():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_read_all_files(fast_reader):
     for testfile in get_testfiles():
         if testfile.get('skip'):
             print('\n\n******** SKIPPING %s' % testfile['name'])
@@ -107,13 +112,18 @@ def test_read_all_files():
             test_opts = testfile['opts'].copy()
             if 'guess' not in test_opts:
                 test_opts['guess'] = guess
+            if 'Reader' in test_opts and 'fast_{0}'.format(test_opts['Reader']._format_name) \
+                in core.FAST_CLASSES: # has fast version
+                if 'Inputter' not in test_opts: # fast reader doesn't allow this
+                    test_opts['fast_reader'] = fast_reader
             table = ascii.read(testfile['name'], **test_opts)
             assert_equal(table.dtype.names, testfile['cols'])
             for colname in table.dtype.names:
                 assert_equal(len(table[colname]), testfile['nrows'])
 
 
-def test_read_all_files_via_table():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_read_all_files_via_table(fast_reader):
     for testfile in get_testfiles():
         if testfile.get('skip'):
             print('\n\n******** SKIPPING %s' % testfile['name'])
@@ -128,6 +138,8 @@ def test_read_all_files_via_table():
                 del test_opts['Reader']
             else:
                 format = 'ascii'
+            if 'fast_{0}'.format(format) in core.FAST_CLASSES:
+                test_opts['fast_reader'] = fast_reader
             table = Table.read(testfile['name'], format=format, **test_opts)
             assert_equal(table.dtype.names, testfile['cols'])
             for colname in table.dtype.names:
@@ -196,25 +208,29 @@ def test_daophot_multiple_aperture():
     assert np.all(table['RAPERT5'] == 23.3)  # assert all the 5th apertures are same 23.3
 
 
-@raises(ascii.InconsistentTableError)
-def test_empty_table_no_header():
-    table = ascii.read('t/no_data_without_header.dat', Reader=ascii.NoHeader,
-                            guess=False)
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_empty_table_no_header(fast_reader):
+    with pytest.raises(ascii.InconsistentTableError):
+        table = ascii.read('t/no_data_without_header.dat', Reader=ascii.NoHeader,
+                            guess=False, fast_reader=fast_reader)
 
 
-@raises(ascii.InconsistentTableError)
-def test_wrong_quote():
-    table = ascii.read('t/simple.txt', guess=False)
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_wrong_quote(fast_reader):
+    with pytest.raises(ascii.InconsistentTableError):
+        table = ascii.read('t/simple.txt', guess=False, fast_reader=fast_reader)
 
 
-@raises(ascii.InconsistentTableError)
-def test_extra_data_col():
-    table = ascii.read('t/bad.txt')
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_extra_data_col(fast_reader):
+    with pytest.raises(ascii.InconsistentTableError):
+        table = ascii.read('t/bad.txt', fast_reader=fast_reader)
 
 
-@raises(ascii.InconsistentTableError)
-def test_extra_data_col2():
-    table = ascii.read('t/simple5.txt', delimiter='|')
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_extra_data_col2(fast_reader):
+    with pytest.raises(ascii.InconsistentTableError):
+        table = ascii.read('t/simple5.txt', delimiter='|', fast_reader=fast_reader)
 
 
 @raises(IOError)
@@ -222,23 +238,28 @@ def test_missing_file():
     table = ascii.read('does_not_exist')
 
 
-def test_set_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_set_names(fast_reader):
     names = ('c1', 'c2', 'c3', 'c4', 'c5', 'c6')
-    data = ascii.read('t/simple3.txt', names=names, delimiter='|')
+    data = ascii.read('t/simple3.txt', names=names, delimiter='|',
+                      fast_reader=fast_reader)
     assert_equal(data.dtype.names, names)
 
 
-def test_set_include_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_set_include_names(fast_reader):
     names = ('c1', 'c2', 'c3', 'c4', 'c5', 'c6')
     include_names = ('c1', 'c3')
     data = ascii.read('t/simple3.txt', names=names, include_names=include_names,
-                           delimiter='|')
+                           delimiter='|', fast_reader=fast_reader)
     assert_equal(data.dtype.names, include_names)
 
 
-def test_set_exclude_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_set_exclude_names(fast_reader):
     exclude_names = ('Y', 'object')
-    data = ascii.read('t/simple3.txt', exclude_names=exclude_names, delimiter='|')
+    data = ascii.read('t/simple3.txt', exclude_names=exclude_names, delimiter='|',
+                      fast_reader=fast_reader)
     assert_equal(data.dtype.names, ('obsid', 'redshift', 'X', 'rad'))
 
 
@@ -311,31 +332,34 @@ def test_set_converters():
     assert_equal(data['p1.gamma'][0], '1.26764544642')
 
 
-def test_from_string():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_from_string(fast_reader):
     f = 't/simple.txt'
     with open(f) as fd:
         table = fd.read()
     testfile = get_testfiles(f)
-    data = ascii.read(table, **testfile['opts'])
+    data = ascii.read(table, fast_reader=fast_reader, **testfile['opts'])
     assert_equal(data.dtype.names, testfile['cols'])
     assert_equal(len(data), testfile['nrows'])
 
 
-def test_from_filelike():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_from_filelike(fast_reader):
     f = 't/simple.txt'
     testfile = get_testfiles(f)
     with open(f, 'rb') as fd:
-        data = ascii.read(fd, **testfile['opts'])
+        data = ascii.read(fd, fast_reader=fast_reader, **testfile['opts'])
     assert_equal(data.dtype.names, testfile['cols'])
     assert_equal(len(data), testfile['nrows'])
 
 
-def test_from_lines():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_from_lines(fast_reader):
     f = 't/simple.txt'
     with open(f) as fd:
         table = fd.readlines()
     testfile = get_testfiles(f)
-    data = ascii.read(table, **testfile['opts'])
+    data = ascii.read(table, fast_reader=fast_reader, **testfile['opts'])
     assert_equal(data.dtype.names, testfile['cols'])
     assert_equal(len(data), testfile['nrows'])
 
@@ -346,35 +370,41 @@ def test_comment_lines():
     assert_equal(table.comment_lines, ['# first comment', '  # second comment'])
 
 
-def test_fill_values():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_fill_values(fast_reader):
     f = 't/fill_values.txt'
     testfile = get_testfiles(f)
-    data = ascii.read(f, fill_values=('a', '1'), **testfile['opts'])
+    data = ascii.read(f, fill_values=('a', '1'), fast_reader=fast_reader,
+                      **testfile['opts'])
     assert_true((data['a'].mask == [False, True]).all())
     assert_true((data['a'] == [1, 1]).all())
     assert_true((data['b'].mask == [False, True]).all())
     assert_true((data['b'] == [2, 1]).all())
 
 
-def test_fill_values_col():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_fill_values_col(fast_reader):
     f = 't/fill_values.txt'
     testfile = get_testfiles(f)
-    data = ascii.read(f, fill_values=('a', '1', 'b'), **testfile['opts'])
+    data = ascii.read(f, fill_values=('a', '1', 'b'), fast_reader=fast_reader,
+                      **testfile['opts'])
     check_fill_values(data)
 
 
-def test_fill_values_include_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_fill_values_include_names(fast_reader):
     f = 't/fill_values.txt'
     testfile = get_testfiles(f)
-    data = ascii.read(f, fill_values=('a', '1'),
+    data = ascii.read(f, fill_values=('a', '1'), fast_reader=fast_reader,
                            fill_include_names = ['b'], **testfile['opts'])
     check_fill_values(data)
 
 
-def test_fill_values_exclude_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_fill_values_exclude_names(fast_reader):
     f = 't/fill_values.txt'
     testfile = get_testfiles(f)
-    data = ascii.read(f, fill_values=('a', '1'),
+    data = ascii.read(f, fill_values=('a', '1'), fast_reader=fast_reader,
                            fill_exclude_names = ['a'], **testfile['opts'])
     check_fill_values(data)
 
@@ -390,11 +420,12 @@ def check_fill_values(data):
     assert_true((data['b'] == [2, 1]).all())
 
 
-def test_fill_values_list():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_fill_values_list(fast_reader):
     f = 't/fill_values.txt'
     testfile = get_testfiles(f)
     data = ascii.read(f, fill_values=[('a', '42'), ('1', '42', 'a')],
-                           **testfile['opts'])
+                      fast_reader=fast_reader, **testfile['opts'])
     data['a'].mask = False  # explicitly unmask for comparison
     assert_true((data['a'] == [42, 42]).all())
 
@@ -445,21 +476,23 @@ def test_set_guess_kwarg():
     assert(len(data) == 1)
 
 
-@raises(ascii.InconsistentTableError)
-def test_read_rdb_wrong_type():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_read_rdb_wrong_type(fast_reader):
     """Read RDB data with inconstent data type (except failure)"""
     table = """col1\tcol2
 N\tN
 1\tHello"""
-    ascii.read(table, Reader=ascii.Rdb)
+    with pytest.raises(ascii.InconsistentTableError):
+        ascii.read(table, Reader=ascii.Rdb, fast_reader=fast_reader)
 
 
-def test_default_missing():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_default_missing(fast_reader):
     """Read a table with empty values and ensure that corresponding entries are masked"""
     table = '\n'.join(['a,b,c,d',
                        '1,3,,',
                        '2, , 4.0 , ss '])
-    dat = ascii.read(table)
+    dat = ascii.read(table, fast_reader=fast_reader)
     assert dat.masked is True
     assert dat.pformat() == [' a   b   c   d ',
                              '--- --- --- ---',
@@ -468,7 +501,7 @@ def test_default_missing():
 
     # Single row table with a single missing element
     table = """ a \n "" """
-    dat = ascii.read(table)
+    dat = ascii.read(table, fast_reader=fast_reader)
     assert dat.pformat() == [' a ',
                              '---',
                              ' --']
@@ -774,12 +807,13 @@ def test_csv_table_read():
     t = ascii.read(lines)
     assert t.colnames == ['a', 'b']
 
-def test_overlapping_names():
+@pytest.mark.parametrize('fast_reader', [True, False, 'force'])
+def test_overlapping_names(fast_reader):
     """
     Check that the names argument list can overlap with the existing column names.
     This tests the issue in #1991.
     """
-    t = ascii.read(['a b', '1 2'], names=['b', 'a'])
+    t = ascii.read(['a b', '1 2'], names=['b', 'a'], fast_reader=fast_reader)
     assert t.colnames == ['b', 'a']
 
 def test_sextractor_units():